/dar, quitar y consultar likes

+++ demo/src/main/java/tp_final_progra3/demo/model/dto/request/UpdateReviewRequestDTO.java
@@ -2,7 +2,7 @@ package tp_final_progra3.demo.model.dto.request;
 
 public record UpdateReviewRequestDTO(
         String contenido,
-        Integer puntuacion,
+        Float puntuacion,
         Boolean contieneSpoilers
 ) {
 }
